requirements.txt
# Meeting Minutes Project Dependencies - Simplified (No Google Cloud Required!)

# Core Framework
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0

# Gemini AI (Only Google Service - Just API Key Needed)
google-generativeai==0.3.2

# Audio Transcription - AssemblyAI (FREE API)
assemblyai==0.33.0

# Audio Processing
pydub==0.25.1

# Document Generation
reportlab==4.0.7
markdown==3.5.1
python-docx==1.1.0

# Utilities
requests==2.31.0
python-dateutil==2.8.2

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-flask==1.3.0

# Development
black==23.12.1
